% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{allgls}
\alias{allgls}
\title{Guideline functions and values}
\format{
A data frame with 6 variables:

\describe{
\item{game}{The name of the game}
\item{release_year}{The year the game was released}
\item{hardware_units_sold}{The amount of hardware units sold}
\item{comment_hardware}{Comment accompanying the amount of hardware units sold}
\item{estimated_gross_revenue}{Estimated gross revenue in US$ with 2019 inflation}
\item{comment_revenue}{Comment accompanying the amount of hardware units sold}
}
}
\usage{
allgls
}
\description{
Source: https://en.wikipedia.org/wiki/Arcade_game#List_of_highest-grossing_games
}
\keyword{datasets}
