% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blmextract.R
\name{blm.extract}
\alias{blm.extract}
\title{BLM extract for Dissolved Copper BC Guidelines}
\usage{
blm.extract(x)
}
\arguments{
\item{x}{data frame - BLM model ouput in the txt format}
}
\value{
data frame in the long format with 4 columns: station and sample (similar to BLM input and output files) and guideline type (either chronic or acute), and guideline values.
}
\description{
Add Text Boxes to Plot when guidelines exceed maximum range of plot.
}
