% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blminput.R
\name{blminput}
\alias{blminput}
\title{Create BLM Input File}
\usage{
blminput(dat1 = NA, lunames = NA)
}
\arguments{
\item{dat1}{Data frame in long format that has been run through sreenCols script.}

\item{lunames}{A character vector of column names}
}
\value{
A dataframe that can be used as input into the BLM software.
}
\description{
Create input file to run BLM to derive copper guideline.
}
