% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenPltcode.R
\name{cenPltcode}
\alias{cenPltcode}
\title{Create Colour and Shapes for Censored Data Plots}
\usage{
cenPltcode(data = NA, statinfo = NA)
}
\arguments{
\item{data}{Data frame with censored data. need to have logical column named "cen" with TRUE for censored values}

\item{statinfo}{data frame with a station column and the corresponding colour (named colour) and shape (named shape). Shape options include: square, triUp, triDown,circle, and diamond.}
}
\value{
A data frame with colours for colour, fill and shape
}
\description{
Add guideline values to table.
}
