% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCenval.R
\name{formatCenval}
\alias{formatCenval}
\title{Format Numbers with Censored values}
\usage{
formatCenval(x, digit = 2, comma = T)
}
\arguments{
\item{x}{character vector with censored data}

\item{digit}{number of significant digits}

\item{comma}{Logical specifiying if a comma should be inserted into numbers greater than 1000.}
}
\description{
Formats numbers to a desired number of digits and keeps "<"'s
}
