% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glAppend.R
\name{glAppend}
\alias{glAppend}
\title{Add Guideline Values to Table}
\usage{
glAppend(dat1 = NA, Location = NA, Type = NA, name = NA)
}
\arguments{
\item{Type}{}
}
\value{
A dataframe with screening values
}
\description{
Add guideline values to table.
}
