% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glTempPh.R
\name{glTempPh}
\alias{glTempPh}
\title{Add Lookup Values for Ammonia}
\usage{
glTempPh(gl = c("CCME", "BC", "PWQO"), temp, pH)
}
\arguments{
\item{gl}{What location will be used}

\item{temp}{field temperature}

\item{value}{Field pH}
}
\value{
A vector with Temperature and pH for screening
}
\description{
Creates ammonia lookup table. Rounding differs for BC, CCME and PWQO.
}
