% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gltextlab.R
\name{gltextlab}
\alias{gltextlab}
\title{Add Guideline Text Boxes to Plot}
\usage{
gltextlab(x, vallist, gllist, gllabel, glcol, Unit, adj = 0.6)
}
\arguments{
\item{x}{data frame in long format}

\item{vallist}{name of columns with plot values and other factors that affect the y maximum (for example normal ranges)}

\item{gllist}{list of column names with guidelines}

\item{gllabel}{list of names of guidelines in same order of gllist}

\item{glcol}{list of colours for text in text box. Same order as gllist}

\item{Unit}{unit of the parameter.}
}
\value{
List of labels to inserted to plot.
}
\description{
Add Text Boxes to Plot when guidelines exceed maximum range of plot.
}
