% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glTextlabel.R
\name{glTextlabel}
\alias{glTextlabel}
\title{Add Guideline Text Boxes to Plot}
\usage{
glTextlabel(gldf, unit)
}
\arguments{
\item{gldf}{data frame in long format with name of gls (pltName),the maximum y values (maxy),the colour of the label and the unit(unit)}

\item{unit}{the unit of the parameter}
}
\value{
Table with labels and colours.
}
\description{
Add Text Boxes to Plot when guidelines exceed maximum range of plot.
}
