% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxCen.R
\name{maxCen}
\alias{maxCen}
\title{Calculate Max Value for Censored Data.}
\usage{
maxCen(vals, format = F, digits = 3)
}
\arguments{
\item{vals}{Character vector with potential censored values.}

\item{format}{Whether to format the output to the desired number of digits.}

\item{digits}{Number of significant digits.}
}
\value{
Maximum value
}
\description{
Calculate and format max value for censored data.  Takes highest detected value. If all values are censored it takes the smalled LRL.
}
