% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meankm.R
\name{meankm}
\alias{meankm}
\title{Calculate Mean for Censored Data.}
\usage{
meankm(vals, digit = 3, format = F)
}
\arguments{
\item{vals}{Character vector with potential censored values.}

\item{format}{Whether to format the output to the desired number of digits.}

\item{digits}{Number of significant digits.}
}
\value{
Mean value.
}
\description{
Calculate and format mean for censored data.
}
