% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minCen.R
\name{minCen}
\alias{minCen}
\title{Calculate Min Value for Censored Data.}
\usage{
minCen(vals, format = F, digits = 3)
}
\arguments{
\item{vals}{Character vector with potential censored values}

\item{format}{Whether to format the output to the desired number of digits}

\item{digits}{Number of significant digits}
}
\value{
Minimum value
}
\description{
Calculate and format min value for censored data.  Retains "<" if smallest value is censored.
}
