% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthaxis.R
\name{monthaxis}
\alias{monthaxis}
\title{Add Years and Months to x-axis labels.}
\usage{
monthaxis(
  x,
  yrstart,
  yrend,
  months = F,
  minusM = 4,
  addM = 0,
  monthsize = 10,
  yrtick = 0.14,
  mntick = 0.06,
  yearsize = 10,
  rotyear = 0
)
}
\arguments{
\item{x}{Plots to change x-axis. The x-axis needs to be dates}

\item{rotyear}{}
}
\value{
Table with labels and colours.
}
\description{
Add Text Boxes to Plot when guidelines exceed maximum range of plot.
}
