% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.standardize.R
\name{name.standardize}
\alias{name.standardize}
\title{standardize parameter names using param table from Teck database}
\usage{
name.standardize(paramdf, namesdf, type = NA)
}
\arguments{
\item{paramdf}{data.frame or list of parameters need to be standardized with  Parameter_Name column}

\item{namesdf}{data.frame with Parameter_Name and Param_Pref columns with Param_Pref is the column for final name and that table is pulled from lu_Param table in Teck database.}
}
\value{
data frame with parameters that do not have the standardized names and data frame with all params name matched up.
}
\description{
standardize parameter names using param table from Teck database.
}
\examples{
 source(paste0(gsub("Documents","Trinity Consultants, Inc/Minnow - General/",path.expand('~')),"Client Data/Teck/Data_Managment/Database/Gather_Data_functions.R"))
}
