% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perkm.R
\name{perkm}
\alias{perkm}
\title{Calculate Percentiles for Censored Data.}
\usage{
perkm(vals, digit = 3, probs, format = T)
}
\arguments{
\item{vals}{Character vector with potential censored values}

\item{probs}{Probabilites to evaluate.}

\item{format}{Whether to format the output to the desired number of digits}

\item{digits}{Number of significant digits}
}
\value{
Percentile values.
}
\description{
Calculate and format percentiles for censored data.
}
