% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdkm.R
\name{sdkm}
\alias{sdkm}
\title{Calculate Standard Deviation for Censored Data.}
\usage{
sdkm(vals, digit = 3, format = T)
}
\arguments{
\item{vals}{Character vector with potential censored values.}

\item{format}{Whether to format the output to the desired number of digits.}

\item{digits}{Number of significant digits.}
}
\value{
Standard deviation.
}
\description{
Calculate and format standard deviation for censored data.
}
