% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitconvert.R
\name{unitconvert}
\alias{unitconvert}
\title{Converts Units for Censored data}
\usage{
unitconvert(x, conval)
}
\arguments{
\item{x}{character vector with censored data}

\item{conval}{unit multiplier}
}
\value{
character vector with converted units and <'s retained
}
\description{
Add guideline values to table.
}
