% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqIndex.R
\name{wqIndex}
\alias{wqIndex}
\title{Calculate CCME Water Quality Index}
\usage{
wqIndex(dat = NA, group = NA, gl = "gl", param = "param")
}
\arguments{
\item{dat}{Data frame with water quality or sediment quality data.}

\item{group}{character vector with column(s) for grouping.}

\item{gl}{name of guideline column}

\item{param}{name of params column}
}
\value{
Original dataframe with sediment quality columns add to the end
}
\description{
Calculate CCME Water Quality Index
}
