% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGLlabels.R
\name{addGLlables}
\alias{addGLlables}
\title{Add Guideline Values to Table}
\usage{
addGLlables(labels = NA, pl = NA, pltcol = NA, labsep = 0.1, xdate = T)
}
\arguments{
\item{labels}{vector of labels to add.}

\item{pl}{plot that labels will be added to.}

\item{pltcol}{Vector of colors for the labels.}

\item{labsep}{Multiplier to separate text labels}

\item{xdate}{Logical stating if x is a date axis or not.}
}
\value{
Plot with labels added.
}
\description{
Add guideline values to table.
}
