% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glAppend.R
\name{glAppend}
\alias{glAppend}
\title{Add Guideline Values to Table}
\usage{
glAppend(dat1 = NA, Location = NA, glType = NA, name = NA)
}
\arguments{
\item{dat1}{Data frame in long format. Should have column with management unit (MU) when using benchmarks.}

\item{Location}{The location of the guidelines. Only BC at the moment.}

\item{Type}{Type of guidelines. Current options include: Long-term chronic, Short-term acute, Teck-BM1,Teck-BM2, and Teck-BM3. The data frame allgls contains all guidelines currently available.}

\item{Name}{The name to use for resulting guideline column. Will result in two columns with "_upper" and "_lower" appended to the name.}
}
\value{
A dataframe with screening values
}
\description{
Add guideline values to table.
}
