% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenCols.R
\name{screenCols}
\alias{screenCols}
\title{Adding Screening Columns to Water Quality Data}
\usage{
screenCols(data = NA, params = NA, value = NA, gls = "BC")
}
\arguments{
\item{data}{Data frame of water quality data. Must have "param" column with the parameter names. Also need "station", and "month" columns.}

\item{params}{Vector of unique water quality parameters in your "param" column.}

\item{value}{Name of values column}

\item{gls}{Which guidelines (BC, CCME)}
}
\value{
A data frame with the appended screening columns.
}
\description{
Adding Screening Columns to Water Quality Data
}
